chore: rename BaseButton to ActionButton
